LayoutTests/fast/scrolling/mac/async-scroll-overflow-rtl-expected.txt


Test scroll over content
PASS overflowScrollEventCount > 0 is true
PASS windowScrollEventCount is 0
PASS successfullyParsed is true

TEST COMPLETE
